consolidated FitnessProfile into UserProfile class. Moved into entity folder.
