+++ Projects/Project1/app/src/main/java/com/example/mcresswell/project01/DashboardActivity.java
@@ -8,10 +8,8 @@ import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 
-import com.example.mcresswell.project01.userProfile.UserProfile;
-import com.example.mcresswell.project01.util.GeocoderLocationUtils;
+import com.example.mcresswell.project01.db.entity.UserProfile;
 import com.example.mcresswell.project01.util.SampleProfileData;
-import com.example.mcresswell.project01.util.ValidationUtils;
 import com.example.mcresswell.project01.weather.WeatherForecast;
 import com.example.mcresswell.project01.weather.WeatherFragment;
