+++ Projects/Project1/app/src/main/java/com/example/mcresswell/project01/FitnessDetailsFragment.java
@@ -14,8 +14,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import com.example.mcresswell.project01.userProfile.UserProfile;
-import com.example.mcresswell.project01.userProfile.UserProfileViewModel;
+import com.example.mcresswell.project01.db.entity.UserProfile;
+import com.example.mcresswell.project01.ViewModels.UserProfileViewModel;
 import com.example.mcresswell.project01.util.Constants;
 import com.example.mcresswell.project01.util.UserProfileUtils;
