+++ Projects/Project1/app/src/main/java/com/example/mcresswell/project01/db/repo/FitnessProfileRepository.java
@@ -1,13 +1,12 @@
-package com.example.mcresswell.project01.db;
+package com.example.mcresswell.project01.db.repo;
 
-import com.example.mcresswell.project01.db.entity.FitnessProfile;
-import com.example.mcresswell.project01.userProfile.UserProfile;
+import com.example.mcresswell.project01.db.entity.UserProfile;
 
 public class FitnessProfileRepository {
     private static final FitnessProfileRepository ourInstance = new FitnessProfileRepository();
 
 
-    private FitnessProfile fitnessProfile;
+    private UserProfile fitnessProfile;
 
     public static FitnessProfileRepository getInstance() {
         return ourInstance;
@@ -19,14 +18,14 @@ public class FitnessProfileRepository {
         //best to get from the Database.
     }
 
-    public FitnessProfile getFitnessProfile() {
+    public UserProfile getFitnessProfile() {
         return fitnessProfile;
     }
 
 
     //TODO: KEEP THIS METHOD. WE'LL NEED THIS METHOD LATER TO BUILD A USER PROFILE OBJECT FROM DATABASE LOOKUP
     public UserProfile userProfileFromFitnessProfile(int fitnessProfileId) {
-        //Retrieve a record from FitnessProfile with the given id
+        //Retrieve a record from UserProfile with the given id
         //Then generate a UserProfile object from the record retrieved
         return null;
     }
