+++ Projects/Project1/app/src/main/java/com/example/mcresswell/project01/util/SampleProfileData.java
@@ -1,6 +1,6 @@
 package com.example.mcresswell.project01.util;
 
-import com.example.mcresswell.project01.userProfile.UserProfile;
+import com.example.mcresswell.project01.db.entity.UserProfile;
 
 import java.util.ArrayList;
 import java.util.Calendar;
